Up our tar game a bit
Drop our lame hacked up tar header implementation for the real thing
from the 'tar' crate.  Reading data directly into it is actually pretty
easy.

Add a reader-side utility class for splitstream.  Also: add a new
SplitStreamData type for inline/external data and port our existing
helper functions to use it.

Add a 'cfsctl ls' command which lists out the content of a tar
splitstream.  So far it just lists files and if they are inline or
external references.  This is the start of support for creating a
dumpfile...

This is some dreadfully ugly code, but it seems to be working.  Probably
some tests would be nice at some point...