oci: clean up tar processing
Simplify the implementation of pax headers, and add support for pax
long names and link targets.

Be a little bit less silly about making copies of borrows when we can
move the original.  Use this to make sure our resulting Entry owns all
its data: we're going to return it soon.