repo: clean up stream-creation operations
Creating a stream is a lot higher-level now: you ask the repo for a
SplitStreamWriter, write to it, then store it back into the repo.

Add a "by-sha256" concept as outlined in #3.  This makes importing
already-import OCI layers instantaneous, even if they were originally
imported by a different user, by a different piece of software, or under
a different reference name.

Closes #3