Add something like a POSIX filesystem
This is something like a POSIX filesystem modelled in rust.  It uses
refcounted references to inodes in order to support hardlinks (but takes
advantage of the fact that hardlinks can't be made to directories).  The
directories are just are just sorted Vecs of entries mapping names to
inodes.

Our cfsctl oci image-create command now plays the layer content into
this filesystem model before printing it out.  The output format is
currently very rough, but it will get cleaned up soon.