dumpfile: overhaul formatting
The Display trait is an awkward fit here because we can't really display
inodes as dumpfile entries without having additional arguments to the
formatting.  We do want the display trait for tar entries, though, I
think?

It might make sense to put this all back to std::io::Write in the
future, but for now this is working.