splitstream: add sha256->fsverity references
Add a header to splitstream to store a series of references from plain
full-content sha256 digests to fsverity digests.  This will be useful
when storing JSON documents as splitstreams, so that they can refer to
other objects in the repository.

Clean up the implementation of SplitStreamReader and make it the only
thing that ever gets returned from Repository::open_stream().  It no
longer does internal buffer allocations when reading tar headers, saving
another ~10% off the total time to print a merged dumpfile.

This is a break in the splitstream file format.  Delete your
repositories and start over!