repository: throw away streams/by-sha256/
There's nothing that we plan to store as a splitstream which we won't
already know the sha256 for.  Drop the by-sha256/ subdirectory and just
call streams by their sha256 content hash directly.

GC remains broken but it should be a bit easier to fix it after this.

This is another "time to delete your repo" change.