oci: create-image now operates on containers
Instead of a list of diffids (ie: layers), the create-image command now
takes a container ID (ie: the sha256 hash of the config, or a reference
to it).

This is a bit unergonomic at present, since the primary thing produced
by the pull operation is a reference to the manifest.  I think we might
want to change that soon, though.