oci: pull image configs, not manifests
Stop storing the manifest locally.  This is getting in the way.

Also: actually use the reference from the user to name the downloaded
container image.

Also: clean up the process of creating images.  SplitStreamReader
implements std::io::Read now, so we can directly parse from it.  We also
do the lookup of the fsverity digests in the splitstream like we should
have all along.  The mkcomposefs call is now in src/image.rs because
it's not OCI-specific.

Now you can do something like:

  cfsctl oci pull docker://docker.io/fedora:40 fedora40
  cfsctl oci create-image refs/fedora40 myimage
  cfsctl mount refs/myimage /mnt