examples: add an example for building an OS image
This builds an OS image from a container, calculating the composefs
fs-verity digest and hardcoding it into the cmdline of a UKI.  When the
UKI boots it finds the composefs repository using the GPT partlabel and
mounts the named image, pivoting the "real" root filesystem to /sysroot
in the usual way.

That's done via a multi-stage container build:

 - build the base image
 - build the kernel image from the base image plus a COMPOSEFS_FSVERITY
   build arg
 - create the final image by copying the kernel UKI into the final image
   and then deleting it again with a whiteout (see below)

The main innovation here is a very dirty little tricky: we've been
looking for a way to include the kernel in the container image and to be
able to refer to the container image from the kernel, which normally
isn't possible for two things that hash each others content.  We break
the cycle by including the kernel in a /composefs-meta directory which
we then record a whiteout for.  After the whiteout, the filesystem is
identical to the base image before the kernel is added, and the fsverity
digest is not impacted by the presence of the kernel.

The kernel itself is, of course, stored in a layer and part of the
container image itself, so it impacts the container ID (and would be
covered by any container signing scheme).

There is a very ugly hack here at present which will not be present in
the final version: we currently need to boot a VM to enable verity on
the composefs objects.  Hopefully mkfs.ext4 will gain support for this
soon.