examples/uki: install systemd-boot normally
I originally tried to put systemd-boot in the ESP but the EFI BIOS
wasn't happy for some reason, so I used -kernel.  Trying again, this is
how `bootctl install` would do it and it seems to work, so let's keep it
that way since it gives us more of a complete/normal image and makes
"bootctl is-installed" report "yes".