cfsctl: add a new 'oci prepare-boot' verb
This prepares a container for boot by creating its composefs and
unpacking its boot resources into a path given by the user (by default
this is /boot).  We can now use this during the image build as well.

This is currently less efficient than it should be: because we don't
have the fs-verity digest here we end up checking the validity of the
entire container, which we don't actually need to do: we're satisfied as
long as the 'meta' layer is correct.  This can be fixed later.

With this change it's now more or less straightforward to deploy a new
container from inside of the running VM.  You just `cfsctl oci pull` it
and then `cfsctl oci prepare-boot`.