src: add selinux re-labelling support
This is a fairly straight-forward operation on the in-memory filesystem
tree.

The actual work done by libselinux here is fairly trivial, and with the
use of the excellent `regex-automata` crate we can mostly do it
ourselves: we just need to consult a list of regexps for each file and
whichever one matches first decides the label.  We can build a DFA to do
all of the "real work" for us in a single step.  There's even probably
more room for improvement: if we figure out a way to deal with cache
invalidation (and path substitutions), we could completely avoid path
building and just pass around state identifiers, but let's keep things
simple for the time being.

It's worth noting that `regex-automata` is fairly complete, but doesn't
support some of the more advanced features of PCRE.  Nobody apparently
uses these features when writing SELinux policy, though, as the
labelling on the booted container is 100% correct (as reported by
`restorecon`).

Still: doing the work ourselves means that we don't need to take a
dependency on an external library which might change its behaviour in
the future.  The labelling sits on our critical "must be deterministic"
path, so this is quite nice.

Closes #29

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>