image: wrap Stat.xattrs in a RefCell
Rc<> doesn't want us to mutate the inner value when it has multiple
outstanding refs (as we have for hardlinks).  That's going to be a
problem for us soon when we want to visit the tree, adding selinux
xattrs.

Use RefCell<> to gain interior mutability on the xattr field.  This is
terribly ad-hoc, but it solves the problem at hand in the least
intrusive way possible.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>