examples/bls: add a new example
Add a proof of concept example for simple container images containing
bootloader spec entries.  This is largely a copy of the uki/ directory
plus a little bit of hacking things around.

There's a lot of things here that need to be fixed: we probably don't
want to use the /composefs-meta/ directory in this case, for example,
but then we need to find a way to avoid the awkwardness of having a
non-empty /boot directory which we'd like to mount over top of.

There are also some sed-based transformations in the build script that
clearly need to be implemented in Rust.

So: it works, but just barely, and not really in the way which it needs
to.  But now that we have uki and bls examples, we can start thinking
about how the `prepare-boot` command should look.  We clearly want to
start drawing more ideas from ostree and bootc for this stuff, though.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>