src: add INLINE_CONTENT_MAX constant
We've been doing this incorrectly by storing all non-empty files
externally.  Add a constant and use it internally.

This means that currently-existing splitstreams need to be regenerated:
they'll have also stored small files as external references.  Add some
extra checks at the splitstream-to-image stage that verifies that the
splitstream has followed the rules correctly: this will help identify
older streams that were built with incorrect rules.  This is another
"delete your respository and start over" change.

We *could* provide bridging code here: in case of too-small external
files in the splitstream, we could read them from the repository and
convert them to inline, but let's save ourselves the bother.

Closes #26

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>