oci: Firm up some questions about the / entry
The / entry doesn't appear in many layer tarballs. Until now, we've
arbitrarily created it root:root, 0755, with mtime set to the epoch.
Let's start thinking about this a bit more rigorously.

Add a doc/oci.md with some of these decisions spelled out more
explicitly.

The upshot: we now use 0555 instead of 0755 and we set the mtime to the
mtime of the newest file in the filesystem (instead of the epoch).

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>