examples: reduce duplication, various cleanups
Create an examples/common/ directory for some of the many copy-pasted
parts between our existing examples.

Rework the fix-verity hack: it's now a proper container image which
contains its product (`fix-verity.efi`) in the image at `/`.  This can
be pulled from `quay.io` now, or (the default) built locally.  This
allows it to share the same caching tricks we use in the other builds,
but also means we should just need to build it a whole lot less often.

Add a `README.md` to `examples/` to explain what each example does.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>