fs: Work around rustix listxattr c_char wonkiness
See https://github.com/bytecodealliance/rustix/pull/1061#issuecomment-2518467534
Basically current rustix 0.38 sometimes may expose the element type as a signed
or unsigned variant, depending on the version of linux-raw-sys and rustix feature
flags.

Add a cast to work around this; if we cared to drop the unsafe we could
use the zerocopy crate.

Signed-off-by: Colin Walters <walters@verbum.org>