oci: add handling for "/" in layer tars
So far we've been testing with local podman builds, which don't produce
these by default, but we need to support these, as they appear in some
bootc images.  It's a bit tricky because our root directory isn't an
Inode, so add a special path.

While we're at it, add some error context to the path where this was
failing before.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>