tree: change approach to mtime of root directory
Instead of hardcoding a bunch of weird values for various parameters in
the root directory of the filesystem, handle the presence of an explicit
'stat' value more explicitly by keeping a separate boolean.

This fixes one of our testcases: in case of an empty filesystem the
sentinel value was slipping through which means we'd end up with an
unusual mtime on our root directory of -1, which showed up in the test
snapshots as 18446744073709551615 (0xffffffffffffffff).

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>